restructure tests to use conftest
