+++ tests/test_diversity_handler.py
@@ -10,25 +10,6 @@ from evident.diversity_handler import (AlphaDiversityHandler,
 import evident.exceptions as exc
 
 
-@pytest.fixture
-def alpha_mock():
-    fname = os.path.join(os.path.dirname(__file__), "data/metadata.tsv")
-    df = pd.read_table(fname, sep="\t", index_col=0)
-    adh = AlphaDiversityHandler(df["faith_pd"], df)
-    return adh
-
-
-@pytest.fixture
-def beta_mock():
-    fname = os.path.join(os.path.dirname(__file__), "data/metadata.tsv")
-    df = pd.read_table(fname, sep="\t", index_col=0)
-    dm_file = os.path.join(os.path.dirname(__file__),
-                           "data/distance_matrix.lsmat.gz")
-    dm = DistanceMatrix.read(dm_file)
-    bdh = BetaDiversityHandler(dm, df)
-    return bdh
-
-
 class TestAlphaDiv:
     def test_init_alpha_div_handler(self):
         fname = os.path.join(os.path.dirname(__file__), "data/metadata.tsv")
