+++ tests/test_exploration.py
@@ -1,20 +1,14 @@
-import os
-
-import pandas as pd
 import pytest
 
-from evident import AlphaDiversityHandler
 from evident import exploration as expl
 
 
-def test_effect_size_by_cat():
-    fname = os.path.join(os.path.dirname(__file__), "data/metadata.tsv")
-    df = pd.read_table(fname, sep="\t", index_col=0)
-    adh = AlphaDiversityHandler(df["faith_pd"], df)
-
+@pytest.mark.parametrize("mock", ["alpha_mock", "beta_mock"])
+def test_effect_size_by_cat(mock, request):
+    dh = request.getfixturevalue(mock)
     df = expl.effect_size_by_category(
-        adh, columns=["perianal_disease", "sex", "classification",
-                      "cd_behavior"]
+        dh, columns=["perianal_disease", "sex", "classification",
+                     "cd_behavior"]
     )
 
     assert ~df.isna().any().any()
@@ -32,14 +26,12 @@ def test_effect_size_by_cat():
     assert df.loc["cd_behavior"]["metric"] == "cohens_f"
 
 
-def test_pairwise_effect_size_by_cat():
-    fname = os.path.join(os.path.dirname(__file__), "data/metadata.tsv")
-    df = pd.read_table(fname, sep="\t", index_col=0)
-    adh = AlphaDiversityHandler(df["faith_pd"], df)
-
+@pytest.mark.parametrize("mock", ["alpha_mock", "beta_mock"])
+def test_pairwise_effect_size_by_cat(mock, request):
+    dh = request.getfixturevalue(mock)
     df = expl.pairwise_effect_size_by_category(
-        adh, columns=["perianal_disease", "sex", "classification",
-                      "cd_behavior"]
+        dh, columns=["perianal_disease", "sex", "classification",
+                     "cd_behavior"]
     )
 
     assert ~df.isna().any().any()
@@ -68,16 +60,15 @@ def test_pairwise_effect_size_by_cat():
         assert grp_counts.loc[grp] == 2
 
 
-def test_no_cols():
-    fname = os.path.join(os.path.dirname(__file__), "data/metadata.tsv")
-    df = pd.read_table(fname, sep="\t", index_col=0)
-    adh = AlphaDiversityHandler(df["faith_pd"], df)
+@pytest.mark.parametrize("mock", ["alpha_mock", "beta_mock"])
+def test_no_cols(mock, request):
+    dh = request.getfixturevalue(mock)
 
     with pytest.raises(ValueError) as exc_info_1:
-        expl.effect_size_by_category(adh)
+        expl.effect_size_by_category(dh)
 
     with pytest.raises(ValueError) as exc_info_2:
-        expl.pairwise_effect_size_by_category(adh)
+        expl.pairwise_effect_size_by_category(dh)
 
     exp_err_msg = "Must provide list of columns!"
     assert str(exc_info_1.value) == str(exc_info_2.value) == exp_err_msg
